// src/api/sys/model/userModel.ts
/**
 * @description: Login interface parameters
 */
export interface LoginParams {
  phone: string;
  password: string;
  gid:number;
}

export interface RoleInfo {
  roleName: string;
  value: string;
}

/**
 * @description: Login interface return value
 */
export interface LoginResultModel {
  userId: string | number;
  token: string;
  role: RoleInfo;
}

/**
 * @description: Get user information return value
 */
export interface GetUserInfoModel {
  roles: RoleInfo[];
  // 用户id
  userId: string | number;
  // 用户名
  username: string;
  // 真实名字
  AccountName: string;
  // 头像
  avatar: string;
  // 介绍
  desc?: string;
}